fix api & add popups
