+++ src/App.js
@@ -1,37 +1,16 @@
 import React from "react";
 import "./App.css";
-import { ShiureiToraCollection } from "./ShiureiToraCollection.js";
-
+import { PopupEdit } from "./PopupEdit.js";
+import { PopupShiuriTora } from "./PopupShiuriTora.js";
+import { FaEdit } from "react-icons/fa";
+import{Slides} from "./Slides"
 function App() {
-  const [count, setCount] = React.useState(1);
+
   const [data, setData] = React.useState({});
   const [time, setTime] = React.useState("");
-  const slides = [
-    {
-      id: 1,
-      numbertext: "1 / 4",
-      src: require("./images/1.jpg"),
-      text: "Caption Text",
-    },
-    {
-      id: 2,
-      numbertext: "2 / 4",
-      src: require("./images/2.jpg"),
-      text: "Caption Text",
-    },
-    {
-      id: 3,
-      numbertext: "3 / 4",
-      src: require("./images/3.jpg"),
-      text: "Caption Text",
-    },
-    {
-      id: 4,
-      numbertext: "4 / 4",
-      src: require("./images/4.jpg"),
-      text: "Caption Text",
-    },
-  ];
+  const [popupEdit ,setPopupEdit] = React.useState(false)
+  const [popupShiuriTora ,setPopupShiuriTora] = React.useState(false)
+ 
   const shiureiToraData = [
     {
       id: 1,
@@ -58,38 +37,29 @@ function App() {
       image: require("./images/Bar_micva.jpeg"),
     },
   ];
-  const dots = [];
 
-  dots.fill("4", 0, 4);
-
-  function handleRightClick() {
-    if (count < slides.length) setCount(count + 1);
-  }
-
-  function handleLeftClick() {
-    if (count > 1) setCount(count - 1);
-  }
-
-  function handleDotClick(e) {
-    setCount(parseInt(e.target.id));
-  }
-
-  const log = () => {
-    console.log("dvvs");
-    return (
-      <div className="popup">
-        <p>hello world</p>
-      </div>
-    );
+  const openPopup = () => {
+    setPopupEdit(false)
+    setPopupShiuriTora(true)
   };
+  const openPopupEdit = () => {
+    setPopupEdit(true)
+    setPopupShiuriTora(false)
+  };
+
+  const closePopup =()=>{
+    setPopupEdit(false)
+    setPopupShiuriTora(false)
+  }
 
   React.useEffect(() => {
     fetch("https://www.hebcal.com/shabbat?cfg=json&geonameid=293397&M=on")
       .then((response) => response.json())
       .then((resp) => {
+        console.log(resp);
         setData(resp);
 
-        const d = new Date(resp.items[1].date);
+        const d = new Date(resp.items[0].date);
         setTime({
           hadlaka: `${d.getHours()}:${d.getMinutes()}`,
           havdala: `${d.getHours() + 1}:${d.getMinutes()}`,
@@ -100,7 +70,9 @@ function App() {
 
   return (
     <div className="App">
-      <ShiureiToraCollection />
+      
+      <PopupEdit openPopup={popupEdit} closePopupfunc={closePopup} />
+      <PopupShiuriTora openPopup={popupShiuriTora} closePopupfunc={closePopup} />
       <header id="header-menu">
         <nav>
           <ul className="header__navbar">
@@ -130,59 +102,8 @@ function App() {
           </ul>
         </nav>
         <div className="header__title">בית כנסת בית אל הרצל יהוד</div>
-        <div className="slideshow-container">
-          {slides.map(
-            (slide) =>
-              slide.id === count && (
-                <div
-                  key={slide.id}
-                  className="mySlides fade"
-                  style={slide.style}
-                >
-                  <div className="numbertext">{slide.numbertext}</div>
-                  <img className="header__img" src={slide.src} alt=""></img>
-                  <div className="text">{slide.text}</div>
-                </div>
-              ),
-          )}
-
-          {/* <!-- Next and previous buttons --> */}
-          {count > 1 && (
-            <button className="prev" onClick={handleLeftClick}>
-              &#10094;
-            </button>
-          )}
-          {count < slides.length && (
-            <button className="next" onClick={handleRightClick}>
-              &#10095;
-            </button>
-          )}
-        </div>
-        <br />
-
-        {/* <!-- The dots/circles --> */}
-        <div style={{ textAlign: "center" }}>
-          <span
-            className={count === 1 ? "dot active" : "dot"}
-            id="1"
-            onClick={handleDotClick}
-          ></span>
-          <span
-            className={count === 2 ? "dot active" : "dot"}
-            id="2"
-            onClick={handleDotClick}
-          ></span>
-          <span
-            className={count === 3 ? "dot active" : "dot"}
-            id="3"
-            onClick={handleDotClick}
-          ></span>
-          <span
-            className={count === 4 ? "dot active" : "dot"}
-            id="4"
-            onClick={handleDotClick}
-          ></span>
-        </div>
+      <Slides/>
+    
       </header>
 
       <main>
@@ -200,16 +121,16 @@ function App() {
               <p>ערבית למצ״ש : 7:50</p>
             </div>
             <div className="parashat-shavoua item1">
-              <p> {data.items ? data.items[2].hebrew : "hee"}</p>
+              <p> {data.items ? data.items[1].hebrew : "hee"}</p>
               <p>
                 {" "}
                 {data.items
-                  ? data.items[1].hebrew + " : " + time.hadlaka
+                  ? data.items[0].hebrew + " : " + time.hadlaka
                   : "hee"}
               </p>
               <p>
                 {data.items
-                  ? data.items[3].hebrew + " : " + time.havdala
+                  ? data.items[2].hebrew + " : " + time.havdala
                   : "hee"}
               </p>
               <p>משיב הרוח </p>
@@ -239,7 +160,9 @@ function App() {
                 <div className="shiurei-tora__content">
                   <h3 className="shiueri-Tora__title">{shiur.title}</h3>
                   <p className="shiueri-Tora__text">{shiur.text}</p>
-                  <button onClick={log}>לצפיה בשיעורים קודמים</button>
+                  <button onClick={openPopup}>לצפיה בשיעורים קודמים</button>
+                  <button className="shiueri-Tora__btn" onClick={openPopupEdit}><FaEdit/></button>
+               
                 </div>
                 <img
                   className="shiueri-Tora__img"
